pyproject.toml
[project]
name = "online-market-place"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.0",
    "google-generativeai>=0.5",
    "jinja2>=3.1",
    "pillow>=10.0",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
]
